Implement canary file generation functionality from contract test inputs files
